Fixing BGP Stub Terminology
The enumeration made it look like "multi-homed" AS was a "stub" AS that also had one or more additional connections. In reality, "stub" is the property of not forwarding traffic, and "multi-homed" is a *different* property of having multiple connections.
This was hinted at in the text when talking about Stub ASes, but not made clear in the definitions.

This commit changes the definitions to hopefully make it more clear.